handle mirrorpeer deletion for client clusters
In a provider-client setup, when we have a mirrorpeer
for both provider and the clients, then if we just want
to delete mirrorpeer for clients it should be allowed.
We should be filtering resources created only for that
client and proceed for deletion accordingly.

And, if there are multiple templates for a client i.e,
multiple drpolicies then until all the drpolicies are deleted
we shouldn't allow the deletion of the mirrorpeer for that
relationship.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>